+++ scripts/make.clj
@@ -130,20 +130,31 @@
 
 (defn pod-test [] (clj! "cli" ["-M:pod-test"]))
 
+(defn ^:private run-task [task-name]
+  (let [lsp-bins (->> [:native :script] (map lsp-bin-filename) distinct)
+        lsp-bins-found (->> lsp-bins (filter fs/exists?) (into #{}))
+        bb (str \" (.get (.command (.info (java.lang.ProcessHandle/current)))) \")]
+    (case (count lsp-bins-found)
+      0 (throw (ex-info "No clojure-lsp executables found." {:searched-for lsp-bins}))
+      1 (p/shell {:dir "cli"} bb task-name (str (fs/path ".." (first lsp-bins-found))))
+      (throw (ex-info "More than one clojure-lsp executables found. Can only work with one."
+                      {:bin-found lsp-bins-found})))))
+
 (defn integration-test
   "Run the integration tests in 'test/integration-test/' using `./clojure-lsp[.bat|.exe]`.
 
   There should only be one clojure-lsp executable found, throws error
   otherwise."
   []
-  (let [lsp-bins (->> [:native :script] (map lsp-bin-filename) distinct)
-        lsp-bins-found (->> lsp-bins (filter fs/exists?) (into #{}))
-        bb (str \" (.get (.command (.info (java.lang.ProcessHandle/current)))) \")]
-    (case (count lsp-bins-found)
-      0 (throw (ex-info "No clojure-lsp executables found." {:searched-for lsp-bins}))
-      1 (p/shell {:dir "cli"} bb "integration-test" (str (fs/path ".." (first lsp-bins-found))))
-      (throw (ex-info "More than one clojure-lsp executables found. Can only work with one."
-                      {:bin-found lsp-bins-found})))))
+  (run-task "integration-test"))
+
+(defn performance-test
+  "Run the performance tests in 'test/integration-test/' using `./clojure-lsp[.bat|.exe]`.
+
+  There should only be one clojure-lsp executable found, throws error
+  otherwise."
+  []
+  (run-task "performance-test"))
 
 (defn ^:private lint
   ([linter] (lint linter {}))